refactor: characters service audit fixes (transaction wrapping + avatar extraction)
- C3: wrap importCharacters in a transaction so settings, character
  upserts, and speaker linking are atomic; fails gracefully on rollback
- CL4: extract toCharacterDetail() helper to deduplicate repeated
  CharacterDetail return shapes across CRUD methods
- CL5: extract avatar management (upload/delete) into dedicated
  characters/avatar.ts module
- linkSpeakersToLines now accepts optional tx parameter for
  transaction-safe caller integration

Closes #333
